otlga/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0
Pillow>=8.3.0

